services/user-management/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
databases[postgresql]==0.8.0
databases[mysql]==0.8.0
databases[sqlite]==0.8.0
asyncpg==0.29.0  # PostgreSQL async driver
aiomysql==0.2.0  # MySQL async driver
aiosqlite==0.19.0  # SQLite async driver

# Authentication & Security
pyjwt==2.8.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0

# Email validation
email-validator==2.1.0

# Monitoring
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.0